+++ tools/netcat-client/create_request-bulk-log.php
@@ -3,11 +3,10 @@
 // Usage: php create_request-log.php connection.json
 
 $connection = json_decode(file_get_contents($argv[1]));
-$token = json_decode(file_get_contents("tokens/sample-app"));
 function f($c) {
  global $token;
  return array(
-        "token" => $token->token,
+        "_t" => time(),
         "datetime" => round(microtime(true) * 1000),
         "logger" => "sample-app",
         "msg" => $c . " Efficient real-time centralized logging server ⚡ ",